requirements.txt
flask
flask-cors
python-dotenv
firecrawl
google-generativeai